requirements.txt
aiogram>=3.0
python-dotenv>=1.0
pandas>=2.2
pyarrow>=15.0
langchain>=0.2
langchain-openai>=0.1
langchain-gigachat>=0.3.10